Revert "Move PHP functions into own files (#21)"
This reverts commit 0f6fb39f5c00c2b44e27a4e2ad4cfe8614a33fd7.
